Refactor extension into modules
