+++ popup.html
@@ -180,7 +180,7 @@
     <button id="debug-toggle" style="margin-top: 10px;">디버그 토글</button>
   </div>
 
-  <script src="popup.js"></script>
+  <script type="module" src="popup.js"></script>
 </body>
 
 </html>
